refactored IC tool to be distributed neatly - part 1
